arari-app/api/requirements.txt
# Backend dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
pytest>=7.4.3
httpx>=0.25.1
pyinstaller>=5.13.2
bcrypt>=4.0.0
black>=23.3.0
ruff>=0.1.0
